requirements.txt
motor == 0.5
crypto == 1.4.1
jws == 0.1.3
pyasn1 == 0.1.7
PyJWT == 1.3.0
python-jwt == 1.0.2
tornado == 4.2.1
